Plumb in BigMesh
This also involved an interface change on Marching: the output functor
now receives local indices (relative to the chunk being passed), and
the output functors handle the offsetting. This is because they were
having to do this anyway when welding, and only SimpleMesh ended up
doing more work. It also removes the need for each output functor to
remember how much was passed to it.