Fix a subtle bug in SplatTreeCL.
I had been lazy about calling writeStart for the coarsest level, where
there was no previous level to chain to. For some reason I thought it
would be okay to just pass the same level back in, but that leads to
garbage values being read in.

Also removed useless context member of the class and deleted the
code that pre-filled the start array with -1 since the start array
is completely written now.