Fix initial ordering of the octree entries list.
Previously each splat was written to 8 entries that were widely
separated, but the effect of this is that the position of the entry in
the final command list depended on which octant it was in. This lead to
invariance problems between octrees. Now, the initial entry list places
each splat as 8 consecutive entries. This is worse for memory
coalescing, so may need to be revisited later (e.g. using __local temp
storage).