Fix deadlock when using multiple devices.
This deadlock was not actually specific to the multi-device case: it can
happen as soon as there are multiple device threads, but is less likely.
The issue was that mesherGroup had fewer total items than the number of
device threads. That makes it possible for some of the device threads to be
waiting to obtain items (and unable to progress until other device
threads complete their push), while those device threads that have items
are unable to proceed because they have work from a newer generation and
need to wait for the older generation device threads to complete.

This thus requires device threads to complete out-of-order, which is
more likely with heterogeneous devices.