Make compilation work on clang++
This mainly involved replacing <tr1/foo> with <boost/tr1/foo.hpp> but
also some corner cases where clang++ was strictly (or possibly just
wrong) in how it applies the standard.

<boost/tr1/cstdint.hpp> doesn't exist (at least in 1.46) and
<boost/tr1/unordered_map.hpp> doesn't work (possibly because it
misdetects libstdc++v3 as not providing the functionality), so those
remain as <tr1/cstdint> and <tr1/unordered_map> for now.