Fix kernel timing with an explicit function call.
The events are now stuffed into a vector and the results are
extracted immediately before writing statistics.

This is still not perfect yet: calls to CLH::enqueueNDRangeKernelSplit
will give incorrect results because it is timing the marker instead of
the actual work. CLH::enqueueNDRangeKernel may also give incorrect
results when there is no work.