Refactor FastPly::Reader
There are now two classes: FastPly::MmapReader and
FastPly::MemoryReader, and a virtual base, FastPly::ReaderBase. This is
quite similar to FastPly::WriterBase.

Additionally, the file handle has been separated from the Reader object,
so that file can be open only as long as it needs to be.

Still to do:
- Add more readers, particularly SyscallReader
- Add command-line option to select the reader class
- More documentation
- Unit tests of a bunch of things
- Refactor the read tests so that they can apply to all the reader classes
- Move MemoryReader into the test code