Make SimpleSplatSet use async reads.
There is a separate thread which does the reads, and queues them to the
main thread. This is currently less effective than the previous approach
of doing the asynchronous stuff in FastBlobSet::computeBlobs, but that
is possibly because the main thread is now being overloaded (CPU use is
close to 100% all the time). That part needs to be parallelised.

Another possible optimization is to make readRaw part of the public
interface for FastPly::Reader::Handle, along with a method to transform
one splat at a time. That would free SimpleFileSet from having to
maintain two buffers of potentially mismatched sizes.