Use libnabo for nearest neighbor search.
This also tidies up the build so that Eigen and libnabo are not required
unless --enable-extras is given, removes the check for CGAL, and
moves the internal knng code and tests back into extras/.

This is about on par with the internal knng code performance-wise, which
suggests that the internal code could benefit from sliding midpoint and
more compact nodes.

It is possible to (rather hackishly) change between libnabo and internal
code using a #define.