Convert FineWorkerGroup to use a CircularBuffer for splats
This is only semi-tested, and it will *definitely* need re-tuning since
the circular buffer is only allocate 1x the --max-host-splats value. It
should probably have a separate command-line option for this.
Additionally, the "spare" setting needs to be increased a lot (and
possibly the WorkQueue class eliminated once everything has been
converted).

This also makes equivalent alterations to mlsgpu-mpi, but they're
untested.