Fix up mlsgpu-mpi to work with new CircularBuffer code for FineBucketGroup
One consequence of this change is that work is solicited before there is
necessarily buffer space. This could easily cause the scatter thread to
stall waiting for receiver to make space available, when other nodes
have the space available. It may be necessary to implement some kind of
negotiation to ensure that the node with the least-full buffer is used
instead of whichever happens to speak up first.