Rewrite passing of data from CoarseBucket to DeviceWorkerGroup
The large host-side buffer was eliminated as it caused bubbles and
spurts in the pipeline. Instead, the coarse bucketing only buffers
up as much as can be delivered to the device in a single shot.
CoarseBucket also takes over responsibility for copying data to the
device, and the buffering in DeviceWorkerGroup is removed.

mlsgpu-mpi is well and truly broken at this point. Also removed
FineBucketGroup since it no longer compiles.

This still needs a lot of work in the command-line passing, including:
- eliminating --mem-host-splats and --mem-host-load
- eliminating --bucket-threads
- parameter validation
- resource usage calculation
- tuning parameters in runstats*